Refactoring of the namespaces
