+++ development/actuators/demo-system-service/src/main/java/eubr/atmosphere/tma/actuator/examples/DataTransferObject.java
@@ -1,4 +1,4 @@
-package eubr.atmosphere.tma.actuator;
+package eubr.atmosphere.tma.actuator.examples;
 
 public class DataTransferObject {
