+++ development/actuators/demo-system-service/src/main/java/eubr/atmosphere/tma/actuator/filters/DecryptFilter.java
@@ -1,4 +1,4 @@
-package eubr.atmosphere.tma.actuator;
+package eubr.atmosphere.tma.actuator.filters;
 
 import java.io.IOException;
